requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
Pillow==10.0.1
tensorflow==2.15.0
keras==2.15.0
kagglehub==0.2.5